refactor(schema): reorganize folder structure into fields/, standard/, runtime/
Separates schema system into logical subfolders:
- fields/: FieldSchema types, factories, and converters (arktype, drizzle)
- standard/: StandardSchema types and JSON Schema conversion
- runtime/: Date utilities, regex, serialization helpers

Renames for clarity:
- fields.ts → fields/factories.ts
- validation.ts → fields/validators.ts
- converters/arktype.ts → fields/to-arktype.ts (etc.)
- standard-schema.ts → standard/types.ts
