+++ packages/epicenter/src/core/schema/fields/factories.ts
@@ -28,13 +28,13 @@
  * ```
  */
 
-import type { DateWithTimezone } from './date-with-timezone';
-import { isDateWithTimezoneString } from './date-with-timezone';
+import type { DateWithTimezone } from '../runtime/date-with-timezone';
+import { isDateWithTimezoneString } from '../runtime/date-with-timezone';
 import type {
 	StandardJSONSchemaV1,
 	StandardSchemaV1,
 	StandardSchemaWithJSONSchema,
-} from './standard-schema';
+} from '../standard/types';
 import type {
 	BooleanFieldSchema,
 	FieldComponent,
