+++ packages/epicenter/src/core/schema/fields/to-arktype.ts
@@ -10,7 +10,7 @@
  * infer ObjectType with composition methods available.
  */
 
-import type { StandardSchemaV1 } from '../standard-schema';
+import type { StandardSchemaV1 } from '../standard/types';
 import { type Type, type } from 'arktype';
 import type { ObjectType } from 'arktype/internal/variants/object.ts';
 import type {
